# docs/swagger.yaml
basePath: /
definitions:
...
  partners.CreatePartnerDto:
    properties:
      email:
        maxLength: 255
        type: string
      name:
        maxLength: 255
        minLength: 2
        type: string
      phone:
        maxLength: 255
        type: string
    required:
    - name
    type: object
...
  partners.PartnerResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      organization_id:
        type: string
      phone:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  partners.UpdatePartnerDto:
    properties:
      email:
        maxLength: 255
        type: string
      id:
        type: string
      name:
        maxLength: 255
        minLength: 2
        type: string
      phone:
        maxLength: 255
        type: string
    required:
    - name
    type: object
...
  utils.EmptyData:
    type: object
  utils.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  utils.PaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  utils.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for Preuvio.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Preuvio API
  version: "1.0"
paths:
...
  /api/partners:
    get:
      description: Retrieves all partners scoped to the authenticated user
      parameters:
      - description: Page (1-based, default 1)
        in: query
        minimum: 1
        name: page
        type: integer
      - description: Limit (default 20, max 50)
        in: query
        maximum: 50
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Partners retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/partners.PartnerResponse'
                  type: array
                meta:
                  $ref: '#/definitions/utils.PaginationMeta'
              type: object
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all partners (paginated)
      tags:
      - partners
    post:
      consumes:
      - application/json
      description: Creates a new partner under an organization
      parameters:
      - description: Partner payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/partners.CreatePartnerDto'
      produces:
      - application/json
      responses:
        "201":
          description: Partner created successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/partners.PartnerResponse'
              type: object
        "400":
          description: Bad request or validation error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new partner
      tags:
      - partners
  /api/partners/{id}:
    delete:
      description: Permanently deletes a partner by ID
      parameters:
      - description: Partner UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Partner deleted successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/utils.EmptyData'
              type: object
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a partner
      tags:
      - partners
    get:
      description: Retrieves a single partner by its UUID
      parameters:
      - description: Partner UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Partner retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/partners.PartnerResponse'
              type: object
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Partner not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a partner by ID
      tags:
      - partners
    put:
      consumes:
      - application/json
      description: Updates an existing partner's details by ID
      parameters:
      - description: Partner UUID
        in: path
        name: id
        required: true
        type: string
      - description: Updated partner payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/partners.UpdatePartnerDto'
      produces:
      - application/json
      responses:
        "200":
          description: Partner updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/partners.PartnerResponse'
              type: object
        "400":
          description: Bad request or validation error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a partner
      tags:
      - partners
...
swagger: "2.0"